examples/search_agent_cli/requirements.txt
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
subconscious-sdk